created a like-button.jsx file and moved over the handClick, useState, and like button to that new jsx file. I made the new like-button.jsx component a client component by adding "use client" to the top of the file and the original page component is automatically a server component because I removed the "useclient" from there. Imported the like-button component to the page.jsx and rendered it in the HomePage component. The button still increments on click except now I have a server and client component separation
