+++ app/page.jsx
@@ -1,6 +1,4 @@
-'use client'
-
-import { useState } from 'react'
+import LikeButton from './like-button.jsx'
 
 function Header({ title }) {
     return <h1>
@@ -12,12 +10,6 @@ function Header({ title }) {
 
 export default function HomePage() {
     const names = [ 'Ada Lovelace', 'Grace Hopper', 'Margaret Hamilton' ]
-
-    const [ likes, setLikes ] = useState(0)
-
-    function handleClick() {
-        setLikes( likes + 1 )
-    }
 
     return (
         <div>
@@ -31,10 +23,8 @@ export default function HomePage() {
                     )
             )}
             </ul>
-
-            <button onClick={handleClick}>
-                Like ({ likes })
-            </button>
+            
+            <LikeButton />
         </div>
     )
 }
